requirements.txt
streamlit
pillow
selenium
webdriver-manager
requests
chromedriver